+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Layer meta data for ADT alpha maps is now exported in JSON format.
 - Textures for ADT alpha maps are now exported to their own relative path if 'Enabled Shared Textures' is enabled.
 - Textures for ADT alpha maps now have resolved names from the listfile, rather than fileDataIDs.
+- Fixed issue in Blender add-on that caused doodads to import slightly offset.
 - Fixed issue that could cause alpha channel to be linked incorrectly in imported material shaders.
 - Fixed issue that preventing pasting into a listfile search under some circumstances.